Send the minimum amount of data for a sample
As of now the whole `sample_t` is sent to userspace which is wasteful
because stacks rarely have the maximum amount of frames we can store.
This commit changes this to send the exact amount of frames that have
been unwound.

Test Plan
=========

Added tests and ran for a a while without issues.